+++ m3cli/plugins/upload-terraform-template.py
@@ -28,18 +28,8 @@ def create_custom_request(request):
     parameters['templateActivity'] = 'UPLOAD'
     variables = parameters.pop('variables', None)
     path_to_file = parameters.pop('variables-file', None)
-    if variables and path_to_file:
-        raise AssertionError(
-            'Cannot use the "--variables" and "--variables-file" parameters '
-            'together'
-        )
-    if path_to_file:
-        if not os.path.isfile(path_to_file):
-            raise AssertionError(f'There is no file by path: "{path_to_file}".')
-        with open(path_to_file, 'r') as file:
-            variables = json.load(file)
-    if variables:
-        parameters['variables'] = handle_variables(variables)
+    if variables or path_to_file:
+        parameters['variables'] = handle_variables(variables, path_to_file)
     return request