+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.116.7] - 2024-06-17
+* Refactor code and handle `Invalid JSON format`. This applies to the following
+commands:
+  * `apply-terraform-template`, `plan-terraform-template`
+  * `upload-terraform-template`, `upload-terraform-template-from-git`
 
 ## [3.116.6] - 2024-05-31
 * Add functionality to use JSON file with variables for commands